+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [these versioning and changelog guidelines][changelog-gui
 ## Unreleased
 ### Bug fixes
 * Fixed an issue with footer help links not expanding to full-width on mobile devices ([#759](https://github.com/Shopify/polaris-react/issues/759))
+* Added breadcrumbs to SkeletonPage ([#985](https://github.com/Shopify/polaris-react/pull/985))
 * Added max-width and auto margin to EmptyState ([#969](https://github.com/Shopify/polaris-react/pull/969))
 * Fixed outline button disabled state styles ([#972](https://github.com/Shopify/polaris-react/pull/972))
 * Fixed Tag so the onRemove function is not imporperly called (thanks [chaddjohnson](https://github.com/chaddjohnson) for the [original issue](https://github.com/Shopify/polaris/issues/235) ) ([#970](https://github.com/Shopify/polaris-react/pull/970))